+++ src/rpmostree-db-builtins.h
@@ -30,4 +30,13 @@ gboolean rpmostree_db_builtin_diff (int argc, char **argv, GCancellable *cancell
 gboolean rpmostree_db_builtin_list (int argc, char **argv, GCancellable *cancellable, GError **error);
 gboolean rpmostree_db_builtin_version (int argc, char **argv, GCancellable *cancellable, GError **error);
 
+gboolean rpmostree_db_option_context_parse (GOptionContext *context,
+                                            const GOptionEntry *main_entries,
+                                            int *argc, char ***argv,
+                                            OstreeRepo **out_repo,
+                                            GFile **out_rpmdbdir,
+                                            gboolean *out_rpmdbdir_is_tmp,
+                                            GCancellable *cancellable,
+                                            GError **error);
+
 G_END_DECLS
