+++ src/rpmostree-rpm-util.h
@@ -18,9 +18,29 @@
  * Boston, MA 02111-1307, USA.
  */
 
+#pragma once
+
+#include <ostree.h>
+
 #include <rpm/rpmlib.h>
 #include <rpm/rpmlog.h>
 #include "rpmostree-util.h"
+
+#include "libgsystem.h"
+
+struct RpmHeaders
+{
+  rpmts ts; /* rpm transaction set the headers belong to */
+  GPtrArray *hs; /* list of rpm header objects from <rpm.h> = Header */
+};
+
+struct RpmHeadersDiff
+{
+  GPtrArray *hs_add; /* list of rpm header objects from <rpm.h> = Header */
+  GPtrArray *hs_del; /* list of rpm header objects from <rpm.h> = Header */
+  GPtrArray *hs_mod_old; /* list of rpm header objects from <rpm.h> = Header */
+  GPtrArray *hs_mod_new; /* list of rpm header objects from <rpm.h> = Header */
+};
 
 struct RpmRevisionData
 {
@@ -34,7 +54,32 @@ rpmhdrs_diff (struct RpmHeaders *l1,
               struct RpmHeaders *l2);
 
 void
-rpmhdrs_diff_prnt_diff (GFile *root1, GFile *root2, struct RpmHeadersDiff *diff);
+rpmhdrs_list (GFile *root,
+              struct RpmHeaders *l1);
+
+char *
+rpmhdrs_rpmdbv (GFile *root,
+                struct RpmHeaders *l1,
+                GCancellable *cancellable,
+                GError **error);
+
+void
+rpmhdrs_diff_prnt_block (GFile *root1,
+                         GFile *root2,
+                         struct RpmHeadersDiff *diff);
+
+void
+rpmhdrs_diff_prnt_diff (GFile *root1,
+                        GFile *root2,
+                        struct RpmHeadersDiff *diff);
+
+struct RpmRevisionData *
+rpmrev_new (OstreeRepo *repo,
+            GFile *rpmdbdir,
+            const char *rev,
+            const GPtrArray *patterns,
+            GCancellable *cancellable,
+            GError **error);
 
 void
 rpmrev_free (struct RpmRevisionData *ptr);
@@ -42,8 +87,3 @@ rpmrev_free (struct RpmRevisionData *ptr);
 GS_DEFINE_CLEANUP_FUNCTION0(struct RpmRevisionData *, _cleanup_rpmrev_free, rpmrev_free);
 #define _cleanup_rpmrev_ __attribute__((cleanup(_cleanup_rpmrev_free)))
 
-struct RpmRevisionData *
-rpmrev_new (OstreeRepo *repo, GFile *rpmdbdir, const char *rev,
-      const GPtrArray *patterns,
-      GCancellable   *cancellable,
-      GError        **error);
