+++ src/rpmostree-util.h
@@ -23,6 +23,7 @@
 #include <gio/gio.h>
 #include <sys/types.h>
 #include <sys/wait.h>
+#include <ostree.h>
 
 void
 _rpmostree_set_error_from_errno (GError    **error,
@@ -55,6 +56,13 @@ _rpmostree_util_update_checksum_from_file (GChecksum    *checksum,
                                            GFile        *src,
                                            GCancellable *cancellable,
                                            GError      **error);
+
+GPtrArray *
+_rpmostree_util_get_commit_hashes (OstreeRepo *repo,
+                                   const char *beg,
+                                   const char *end,
+                                   GCancellable *cancellable,
+                                   GError **error);
 
 gboolean
 _rpmostree_sync_wait_on_pid (pid_t          pid,
